+++ rewatch/src/build.rs
@@ -80,7 +80,7 @@ pub fn get_compiler_args(rescript_file_path: &Path) -> Result<String> {
         &project_context.current_config,
         relative_filename,
         &contents,
-    );
+    )?;
     let is_interface = filename.to_string_lossy().ends_with('i');
     let has_interface = if is_interface {
         true